BUILD: opentracing: fixed build when using pkg-config utility
In case the OpenTracing C Wrapper library was installed as part of the
system (ie in a directory that pkg-config considers part of the system),
HAProxy building was not possible because in that case pkg-config did
not set the value of the OT_CFLAGS variable in the addon Makefile.

This resolves GitHub issue #1323.

(cherry picked from commit 3571111de3c21bb6b88cb3637a6f5850cb810de7)
Signed-off-by: Willy Tarreau <w@1wt.eu>