BUG/MINOR: buffer: fix buffer_dump() formatting
The formatting of the buffer_dump() output must be calculated using the
relative counter, not the absolute one, or everything will be broken if
the <from> variable is not a multiple of 16.

Could be backported in all maintained versions.

(cherry picked from commit 7e7765a4515bc23d536e2840c5bbdf18ffca6d45)
Signed-off-by: Willy Tarreau <w@1wt.eu>