BUG/MINOR: tcpcheck: Properly detect pending HTTP data in output buffer
In tcpcheck_eval_send(), the condition to detect there are still pending
data in the output buffer is buggy. Presence of raw data must be tested for
TCP connection only. But a condition on the connection was missing to be
sure it is not an HTX connection.

This patch must be backported as far as 2.2.

(cherry picked from commit 47bfd7b9b78c71ffa08d65f0cef475f5d2ae7b80)
Signed-off-by: Willy Tarreau <w@1wt.eu>